Feat: add blueprints management system
