+++ market/__init__.py
@@ -3,7 +3,7 @@ from flask import Flask, render_template
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
 from flask_sqlalchemy import SQLAlchemy
-from market import routes
+from market.blueprints import blueprints
 from market.constants import FLASK_SECRET_KEY, SQLALCHEMY_DATABASE_URI
 
 
@@ -17,3 +17,6 @@ db = SQLAlchemy(app)
 bcrypt = Bcrypt(app)
 login_manager = LoginManager(app)
 login_manager.login_view = "login_page"
+
+for blueprint in blueprints:
+    app.register_blueprint(blueprint, url_prefix="/{}".format(blueprint.name))
