+++ market/blueprints/__init__.py
@@ -1,0 +1,18 @@
+from .auth import auth
+from .baby_products import baby_products
+from .beverages import beverages
+from .cart import cart
+from .food import food
+from .pet_products import pet_products
+from .snacks import snacks
+
+
+blueprints = [
+    auth,
+    baby_products,
+    beverages,
+    cart,
+    food,
+    pet_products,
+    snacks,
+]
